Code Refactor after finding plugin by same name
FireLess is a firebug extension targeting only the JS port of the less
framework.  I didn't find this in my initial research, and it is only
targeting the js version of the {less} framework.
https://addons.mozilla.org/en-US/firefox/addon/fireless/?src=search

My plugin is only targeting the dotLess port so far, so the code was
refactored to remove the name conflict and to reflect the current state
of the plugin.